+++ packages/teams-js/src/internal/videoPerformanceStatistics.ts
